Tell both players (sender...
...and receiver) that the request sent timed out.
Tested with MT/MTG 5.0.1 and works fine.

The functions "clear_tpr_list" and "clear_tphr_list" are not used due to a crash while sending the message to both players.
Sorry for any inconveniences. I'll (soon?) probably create a fix for this.
If anyone has a fix for this, it would be greatly appreciated. You may comment on my PR.